% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coverageObj.scopeDemo}
\alias{coverageObj.scopeDemo}
\title{Pre-stored coverageObj.scope data for demonstration purposes}
\format{Pre-computed using whole genome sequencing data of
three single cells from 10X Genomics Single-Cell CNV solution}
\usage{
coverageObj.scopeDemo
}
\description{
Pre-stored coverageObj.scope data for demonstration purposes
}
\keyword{datasets}
