% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gc.R
\name{get_gc}
\alias{get_gc}
\title{Compute GC content}
\usage{
get_gc(ref, hgref = NULL)
}
\arguments{
\item{ref}{GRanges object returned from \code{get_bam_bed}}

\item{hgref}{reference genome. This should be either 'hg19' or 'hg38'.}
}
\value{
\item{gc}{Vector of GC content for each bin/target}
}
\description{
Compute GC content for each bin
}
\examples{
library(WGSmapp)
bamfolder <- system.file('extdata', package = 'WGSmapp')
bamFile <- list.files(bamfolder, pattern = '*.dedup.bam$')
bamdir <- file.path(bamfolder, bamFile)
sampname_raw <- sapply(strsplit(bamFile, '.', fixed = TRUE), '[', 1)
bambedObj <- get_bam_bed(bamdir = bamdir,
                            sampname = sampname_raw)
bamdir <- bambedObj$bamdir
sampname_raw <- bambedObj$sampname
ref_raw <- bambedObj$ref

gc <- get_gc(ref_raw)

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
