% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gini.R
\name{get_gini}
\alias{get_gini}
\title{Compute Gini coefficients for single cells}
\usage{
get_gini(Y)
}
\arguments{
\item{Y}{raw read depth matrix after quality control procedure}
}
\value{
\item{Gini}{Vector of Gini coefficients for single cells
    from scDNA-seq}
}
\description{
Gini index is defined as two times the area
between the Lorenz curve and the diagonal.
}
\examples{
Gini <- get_gini(Y_sim)

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
