% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mapp.R
\name{get_mapp}
\alias{get_mapp}
\title{Compute mappability}
\usage{
get_mapp(ref, hgref = NULL)
}
\arguments{
\item{ref}{GRanges object returned from \code{get_bam_bed}}

\item{hgref}{reference genome. This should be either 'hg19' or 'hg38'.}
}
\value{
\item{mapp}{Vector of mappability for each bin/target}
}
\description{
Compute mappability for each bin. Note that scDNA
sequencing is whole-genome amplification and the mappability
score is essential to determine variable binning method.
Mappability track for 100-mers on the GRCh37/hg19 human
reference genome from ENCODE is pre-saved. Compute the mean
of mappability scores that overlapped reads map to bins,
weighted by the width of mappability tracks on the genome
reference. Use liftOver utility to calculate mappability
for hg38, which is pre-saved as well.
}
\examples{
library(WGSmapp)
bamfolder <- system.file('extdata', package = 'WGSmapp')
bamFile <- list.files(bamfolder, pattern = '*.dedup.bam$')
bamdir <- file.path(bamfolder, bamFile)
sampname_raw <- sapply(strsplit(bamFile, '.', fixed = TRUE), '[', 1)
bambedObj <- get_bam_bed(bamdir = bamdir,
                            sampname = sampname_raw)
bamdir <- bambedObj$bamdir
sampname_raw <- bambedObj$sampname
ref_raw <- bambedObj$ref

data('mapp_hg19')
mapp <- get_mapp(ref_raw)

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
