% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcoverage.scDNA.R
\name{getcoverage.scDNA}
\alias{getcoverage.scDNA}
\title{Get read coverage from single-cell DNA sequencing}
\usage{
getcoverage.scDNA(bambedObj, mapqthres, seq, hgref = "hg19")
}
\arguments{
\item{bambedObj}{object returned from \code{getbambed_scope}}

\item{mapqthres}{mapping quality threshold of reads}

\item{seq}{the sequencing method to be used. This should be either
'paired-end' or 'single-end'}

\item{hgref}{human reference genome. This should be either 'hg19' or 'hg38'.}
}
\value{
\item{Y}{Read depth matrix}
}
\description{
Get read coverage for each genomic bin across all single
    cells from scDNA-seq. Blacklist regions, such as segmental duplication 
    regions and gaps near telomeres/centromeres will be masked prior to
    getting coverage.
}
\examples{
library(WGSmapp)
bamfolder <- system.file('extdata', package = 'WGSmapp')
bamFile <- list.files(bamfolder, pattern = '*.dedup.bam$')
bamdir <- file.path(bamfolder, bamFile)
sampname_raw = sapply(strsplit(bamFile, '.', fixed = TRUE), '[', 1)
bambedObj <- getbambed_scope(bamdir = bamdir,
                            sampname = sampname_raw)

# Getting raw read depth
coverageObj <- getcoverage.scDNA(bambedObj,
                                mapqthres = 40,
                                seq='paired-end')
Y_raw <- coverageObj$Y

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
