% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgc_scope.R
\name{getgc_scope}
\alias{getgc_scope}
\title{Compute GC content}
\usage{
getgc_scope(ref, genome = NULL)
}
\arguments{
\item{ref}{GRanges object returned from \code{getbambed_scope}}

\item{genome}{by default, \code{genome = BSgenome.Hsapiens.UCSC.hg19}.
To calculate mappability for hg38, specify
\code{genome = BSgenome.Hsapiens.UCSC.hg38}}
}
\value{
\item{gc}{Vector of GC content for each bin/target}
}
\description{
Compute GC content for each bin
}
\examples{
library(WGSmapp)
bedFile <- system.file('extdata',
                        'scWGA500kbsort.bed',
                        package = 'SCOPE')
bamfolder <- system.file('extdata', package = 'WGSmapp')
bamFile <- list.files(bamfolder, pattern = '*.dedup.bam$')
bamdir <- file.path(bamfolder, bamFile)
sampname_raw = sapply(strsplit(bamFile, '.', fixed = TRUE), '[', 1)
bambedObj <- getbambed_scope(bamdir = bamdir,
                            bedFile = bedFile,
                            sampname = sampname_raw)
bamdir <- bambedObj$bamdir
sampname_raw <- bambedObj$sampname
ref_raw <- bambedObj$ref

data('mapp_hg19')
gc <- getgc_scope(ref_raw, genome = BSgenome.Hsapiens.UCSC.hg19)

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
