% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{internal_load_data}
\alias{internal_load_data}
\title{internal_load_data}
\usage{
internal_load_data()
}
\value{
A data.frame
}
\description{
internal method to load data
}
\details{
This method loads a big dataset that takes a while to load
}
\author{
Thomas Rosendal
}

