% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data()
}
\value{
A data.frame
}
\description{
Exported method to load data
}
\details{
This method loads a dataset only when it has not already been
loaded
}
\author{
Thomas Rosendal
}

