% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{load_data_closure}
\alias{load_data_closure}
\title{load_data_closure}
\usage{
load_data_closure()
}
\value{
A function
}
\description{
Build the closure
}
\details{
This returns a function with an environment that contains the
object x = NULL.
}
\author{
Thomas Rosendal
}

