% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{normObj.scopeDemo}
\alias{normObj.scopeDemo}
\title{Pre-stored normObj.scope data for demonstration purposes}
\format{Pre-computed by SCOPE using pre-stored data \code{Y_sim}}
\usage{
normObj.scopeDemo
}
\description{
Pre-stored normObj.scope data for demonstration purposes
}
\keyword{datasets}
