% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QCmetric.scopeDemo}
\alias{QCmetric.scopeDemo}
\title{Pre-stored QCmetric data for demonstration purposes}
\format{Pre-computed using whole genome sequencing data of
three single cells from 10X Genomics Single-Cell CNV solution}
\usage{
QCmetric.scopeDemo
}
\description{
Pre-stored QCmetric data for demonstration purposes
}
\keyword{datasets}
