% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref.scopeDemo}
\alias{ref.scopeDemo}
\title{Pre-stored 500kb-size reference genome for demonstration purposes}
\format{Pre-computed using whole genome sequencing data
with GC content and mappability scores}
\usage{
ref.scopeDemo
}
\description{
Pre-stored 500kb-size reference genome for demonstration purposes
}
\keyword{datasets}
