% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_sim}
\alias{ref_sim}
\title{A reference genome in the toy dataset}
\format{A GRanges object with 1544 bins and 1 metadata column of GC content}
\usage{
ref_sim
}
\description{
A reference genome in the toy dataset
}
\keyword{datasets}
