% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y_sim}
\alias{Y_sim}
\title{A read count matrix in the toy dataset}
\format{A read count matrix with 1544 bins and 39 cells}
\usage{
Y_sim
}
\description{
A read count matrix in the toy dataset
}
\keyword{datasets}
