% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbambed_scope.R
\name{getbambed_scope}
\alias{getbambed_scope}
\title{Get bam file directories, sample names, and whole genomic bins}
\usage{
getbambed_scope(bamdir, bedFile, sampname)
}
\arguments{
\item{bamdir}{vector of the directory of a bam file. Should be in the same
order as sample names in \code{sampname}.}

\item{bedFile}{path to the whole genome bed file}

\item{sampname}{vector of sample names. Should be in the same order as bam
directories in \code{bamdir}.}
}
\value{
A list with components
    \item{bamdir}{A vector of bam directories}
    \item{sampname}{A vector of sample names}
    \item{ref}{A GRanges object specifying whole genomic bin positions}
}
\description{
Get bam file directories, sample names, and whole genomic
bins from .bed file
}
\examples{
library(WGSmapp)
bedFile <- system.file('extdata', 'scWGA500kbsort.bed', package = 'SCOPE')
bamfolder <- system.file('extdata', package = 'WGSmapp')
bamFile <- list.files(bamfolder, pattern = '*.dedup.bam$')
bamdir <- file.path(bamfolder, bamFile)
sampname_raw = sapply(strsplit(bamFile, '.', fixed = TRUE), '[', 1)
bambedObj <- getbambed_scope(bamdir = bamdir, bedFile = bedFile,
                            sampname = sampname_raw)
bamdir <- bambedObj$bamdir
sampname_raw <- bambedObj$sampname
ref_raw <- bambedObj$ref

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
